chatbot/requirements.txt
flask
pymongo
rapidfuzz
python-dotenv
